Word search finder.
Used a while loop with an if statement to iterate through each word in the text document, calling a function that will return true or false to whether the word read is found within the text. This will increase the match_count variable as well as the word_count in order to move to the next string variable.